+++ packages/datagrid/src/types/index.ts
@@ -1,12 +1,43 @@
+import { ButtonHTMLAttributes, DOMAttributes, HTMLAttributes } from 'react';
+
+import { ButtonIcon } from '@talend/design-system';
+
 import { QUALITY_EMPTY_KEY, QUALITY_INVALID_KEY, QUALITY_VALID_KEY } from '../constants';
 
-export interface QualityEntry {
-	percentage: number;
-	total: number;
+export interface Quality {
+	[QUALITY_INVALID_KEY]: number;
+	[QUALITY_EMPTY_KEY]: number;
+	[QUALITY_VALID_KEY]: number;
 }
 
-export interface Quality {
-	[QUALITY_INVALID_KEY]: QualityEntry;
-	[QUALITY_EMPTY_KEY]: QualityEntry;
-	[QUALITY_VALID_KEY]: QualityEntry;
+export interface HeaderComponentParams {
+	description?: string;
+	typeLabel?: string;
+	semanticTypeLabel?: string;
+	required?: boolean;
+	quality?: Quality;
+	isLoading?: boolean;
+	draftType?: string;
+	menuProps?: Omit<Parameters<typeof ButtonIcon>[0], 'icon' | 'size'> & {
+		'data-feature'?: string;
+	};
+	qualityBarProps?: any;
+}
+
+export interface AgGridCellValue {
+	name: string;
+	value: string;
+}
+
+export interface AgCellEditorRendererPropTypes {
+	colDef: {
+		cellEditorParams?: Record<string, any>;
+		cellEditorPopup?: boolean;
+		domain: string;
+	};
+	eGridCell: HTMLDivElement;
+	stopEditing: (variable?: boolean) => void;
+	value: {
+		value: string;
+	};
 }