+++ tools/scripts-config-storybook-lib/.storybook-templates/preview.js
@@ -45,15 +45,22 @@ const defaultPreview = {
 		mswDecorator,
 		(Story, context) => {
 			i18n.changeLanguage(context.globals && context.globals.locale);
-			return React.createElement(React.Suspense, { fallback: null }, [
-					React.createElement(I18nextProvider, { i18n: i18n, key: 'i18n' }, [
-						React.createElement(IconsProvider, {
-							bundles: ['https://unpkg.com/@talend/icons/dist/svg-bundle/all.svg'],
-							key: 'icons'
-						}),
-						React.createElement(Story, {...context, key: 'story'}),
-					])
-			]);
+			return React.createElement(React.Suspense, { fallback: null },
+				React.createElement(I18nextProvider, { i18n: i18n, key: 'i18n' },
+					React.createElement(Story, {...context, key: 'story'})
+				)
+			);
+		},
+		(Story, context) => {
+			const storyElement = React.createElement(Story, {...context, key: 'story'});
+			return [
+				React.createElement(IconsProvider, {
+					key: 'icons-provider-decorator'
+				}),
+				React.createElement(ThemeProvider, {
+					key: 'theme-provider-decorator'
+				}, storyElement)
+			];
 		},
 		cmfDecorator
 	].filter(Boolean),