+++ tools/scripts-config-storybook-lib/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @talend/scripts-config-storybook-lib
 
+## 1.2.2
+
+### Patch Changes
+
+- 2f40e092e: fix: remove bundles props of IconsProvider
+
+## 1.2.1
+
+### Patch Changes
+
+- 29302bc1f: chore(storybook): without i18n config
+
 ## 1.2.0
 
 ### Minor Changes